Expand to running 3 gunicorn workers
For the rationale of why we chose 3, see https://github.com/canonical-web-and-design/ubuntu.com/pull/11212.

Note that this is somewhat risky. When we did this on ubuntu.com, it locked up the database
and had to [be reverted](https://github.com/canonical-web-and-design/ubuntu.com/pull/11219).

This appears to be because we had used up all 100 connections to the database. Our understanding
is that this is because each worker creates a separate database connection and ubuntu.com
consists of 35 pods in production and 19 pods in staging (54 total). So when we released the change
to expand each of these pods from running 1 worker to 3 workers, we would have added
162 potential extra database connections.

By contrast, we only run 5 pods of this project in production, and 3 in staging (8 total).
So if we expand our workers by a factor of three, this should only add a maximum of 16
extra database connections, which it should  be able to handle.

So hopefully this change is safe.