Tidy up schema validation of release_codenames and status_statuses
We were going out of our way to make sure we pulled the `release_codenames`
and `status_statuses` from the database as early as possible so we could then
use them in the creation of Schema objects. This structure causes a lot of problems,
because it means the database absolutely has to be set up and ready before other things,
which makes it matter a whole lot the exact order in which we import things, which is
pretty limiting.

So here I'm making it so any logic that compares release codenames to the ones
pulled from the database will happen only when a view is called, not before.