Switch code to using flask-sqlalchemy and flask-migrate
There's quite a backstory.

(tl;dr: flask-sqlalchemy is both more common and more straightforward than
using SQLAlchemy core)

For the last couple of weeks I've been trying to rewrite the tests to use a real
database rather than alchemy-mock. That will be in an upcoming PR, and I'll
leave it until then to explain the rationale.

However, as I tried to implement this, I ran into so many problems with trying
to reset the test data in the DB for each test. This is because clearing out or
resetting the database without the db_session getting confused turns out to be
surprisingly hard.

In the course of reading various message boards and stackoverflow answers to try
to solve my problems, I came to realise that there are basically two different ways
that people tend to set up SQLAlchemy:

1. Following [the SQLAlchemy docs][1], where:
    - you manually create an `engine` object, a `db_session` object from the engine
        and a `Base = declarative_base()` object
    - all model objects inherit from Base
    - model objects themselves are dumb - to query the database you do:
        `db_session.query(Model).get()`
2. Using [flask-sqlalchemy][2], where:
    - You create a single `db = SQLAlchemy()` object, and then bind it to your app
    - `db.session` and `db.engine` are available directly on the `db` object
    - all models inherit from `db.Model`
    - model object contain methods for interacting with the DB, e.g.:
        `Model.query.get()`

I actually came across the second structure much more often. It seems to be a lot
more common. Also, the examples written for the second flask-sqlalchemy scenario
were always easier to follow.

Then a couple of days ago when I *still* hadn't got my DB tests to work properly,
@jpmartinspt pointed out to me that the structure I was trying to implement is
actually exactly the one recommended in an older version of the
[flask-testing documentation][3], and the example provided there uses
flask-sqlalchemy. So it looks kinda like it's going to make my life a lot
easier.

flask-sqlalchemy does have drawbacks though, because it needs to be bound to
an application, and then needs to be run within that application's context.

This meant that I had to still create the `db` object in a very short `database.py`
file just so I could import it into `models.py` without a circular dependency complaint.

It also means that to manipulate objects in the python interpreter, I have to first
call `app.app_context().push()` to ensure the `db` is running within the right context.

1: https://docs.sqlalchemy.org/en/14/orm/session_basics.html
2: https://flask-sqlalchemy.palletsprojects.com/en/2.x/
3: https://flask-testing.readthedocs.io/en/v0.4/#testing-with-sqlalchemy